fix: Refactor fee engine tests and remove unused PriorityOrderer class
